# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KuruOps API
  version: "0.1.0"
  description: >
    REST API for KuruOps, a single-tenant SOC alert/incident-management
    platform (Go backend, PostgreSQL, React frontend). This document covers
    every route mounted by internal/httpserver/router.go: the unauthenticated
    /auth/** login flows, /healthz and /metrics, and the JWT-protected
    /api/v1/** surface (alerts, incidents, playbooks, tags, uploads,
    dashboard, live event stream, and admin-only /settings/** configuration).
    See the repository root README for product context, deployment, and the
    default-admin-first-login flow; see TROUBLESHOOTING.md for operational
    issues. This spec favors completeness of shape (paths, params, request/
    response bodies, status codes) over exhaustive field-level validation
    constraints.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/incidents/{id}/roles/{role}:
    put:
      tags: [Incidents]
      summary: Assign users to a NIST 800-61 team role
      description: >
        Role comes from the path (one of commander, incident_handler,
        communications_lead, privacy_officer, technical_lead). commander and
        technical_lead are single-assignee; an unknown role is a 400, not a
        routing 404.
      parameters:
        - $ref: '#/components/parameters/IncidentId'
        - name: role
          in: path
          required: true
          schema: { $ref: '#/components/schemas/IncidentRole' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userIds:
                  type: array
                  items: { type: string, format: uuid }
              required: [userIds]
      responses:
        "204":
          description: Role assignment updated
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/settings/llm-providers:
    get:
      tags: [Settings - LLM Providers]
      summary: List LLM providers
      responses:
        "200":
          description: LLM providers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LLMProvider' }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Settings - LLM Providers]
      summary: Create an LLM provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveLLMProviderRequest'
      responses:
        "201":
          description: Provider created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LLMProvider' }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v1/settings/llm-providers/{id}:
    put:
      tags: [Settings - LLM Providers]
      summary: Update an LLM provider
      description: Empty apiKey means "keep the existing key".
      parameters:
        - $ref: '#/components/parameters/LLMProviderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveLLMProviderRequest'
      responses:
        "200":
          description: Provider updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LLMProvider' }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Settings - LLM Providers]
      summary: Delete an LLM provider
      parameters:
        - $ref: '#/components/parameters/LLMProviderId'
      responses:
        "204":
          description: Deleted
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v1/settings/llm-providers/{id}/default:
    post:
      tags: [Settings - LLM Providers]
      summary: Set this provider as the tenant default
      parameters:
        - $ref: '#/components/parameters/LLMProviderId'
      responses:
        "204":
          description: Default set
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'

  # ----------------------------------------------------- Settings: MCP Servers
...
  /api/v1/settings/roles:
    get:
      tags: [Settings - Roles]
      summary: List roles
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Role' }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Settings - Roles]
      summary: Create a role
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SaveRoleRequest' }
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Role' }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /api/v1/settings/roles/{id}:
    put:
      tags: [Settings - Roles]
      summary: Update a role
      parameters:
        - $ref: '#/components/parameters/RoleId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SaveRoleRequest' }
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Role' }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
    delete:
      tags: [Settings - Roles]
      summary: Delete a role
      parameters:
        - $ref: '#/components/parameters/RoleId'
      responses:
        "204":
          description: Deleted
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          description: Role is still assigned to one or more users -- reassign them first
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ---------------------------------------------- Settings: Identity Providers
...
  /api/v1/settings/retention:
    get:
      tags: [Settings - Retention]
      summary: Get the tenant's data retention policy
      description: >
        Unlike most Settings GETs, this never returns null: retention is on
        by default (18 months per resource type), so an unconfigured tenant
        still gets real numbers back, with configured=false marking them as
        the synthesized default rather than an explicit save.
      responses:
        "200":
          description: Retention config -- always non-null
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetentionConfig'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Settings - Retention]
      summary: Set the tenant's data retention policy
      description: >
        Lowering either value below an existing closed alert/incident's age
        causes it to be permanently deleted on the next hourly sweep --
        evidence in blob storage is never touched, only the tool's own
        record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                alertRetentionMonths: { type: integer, minimum: 0 }
                incidentRetentionMonths: { type: integer, minimum: 0 }
              required: [alertRetentionMonths, incidentRetentionMonths]
      responses:
        "204":
          description: Saved
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'

  # ----------------------------------------------------------- Settings: Slack
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Required on every /api/v1/** route. /auth/** is intentionally
        unauthenticated (that's how a session is obtained in the first
        place), and /healthz and /metrics are unauthenticated for
        infrastructure health checks and scraping. A caller whose
        mustChangePassword flag is still set can reach nothing under
        /api/v1/** except POST /api/v1/account/change-password (see
        RequirePasswordChanged), receiving 403 on everything else until the
        password is changed.
  parameters:
...
    IncidentId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    LLMProviderId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    RoleId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Malformed request body, path, or query parameter
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Unauthorized:
      description: Missing, malformed, or expired bearer token
      content:
        text/plain:
          schema:
            type: string
            description: >-
              Plain-text error message, issued by the auth middleware via
              http.Error -- not the handler-level JSON error shape used
              elsewhere in this API.
    Forbidden:
      description: Authenticated, but the caller's role/resourceAccess doesn't permit this action (or the account still must change its password)
      content:
        text/plain:
          schema:
            type: string
            description: >-
              Plain-text error message, issued by the auth middleware via
              http.Error -- not the handler-level JSON error shape used
              elsewhere in this API.
...
    InternalError:
      description: Unexpected server-side failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      description: Standard handler-level error shape, written by handlers.writeError.
      properties:
        error:
          type: string
      required: [error]
...
    IncidentRole:
      type: string
      enum: [commander, technical_lead, incident_handler, communications_lead, privacy_officer]
...
    ResourceAccess:
      type: array
      description: Capability set; any combination of the three values.
      items:
        type: string
        enum: [alerts, incidents, followup]
...
    Role:
      type: object
      description: >-
        A named, reusable access bundle -- users and IdP group mappings are
        assigned to one via roleId, rather than carrying
        isAdmin/resourceAccess/allowedTags individually.
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        name: { type: string }
        isAdmin: { type: boolean }
        resourceAccess: { $ref: '#/components/schemas/ResourceAccess' }
        allowedTags:
          type: array
          items: { type: string }
          description: Empty = unrestricted, sees all tags.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, tenantId, name, isAdmin, resourceAccess, allowedTags, createdAt, updatedAt]
    SaveRoleRequest:
      type: object
      properties:
        name: { type: string }
        isAdmin: { type: boolean }
        resourceAccess: { $ref: '#/components/schemas/ResourceAccess' }
        allowedTags:
          type: array
          items: { type: string }
      required: [name, isAdmin, resourceAccess, allowedTags]
...
    LLMProvider:
      type: object
      description: APIKeySecretRef is never serialized; the key itself is never stored in this table.
      properties:
        id: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string }
        baseUrl: { type: string, nullable: true }
        model: { type: string }
        isDefault: { type: boolean }
        autoAnalyzeAllAlerts:
          type: boolean
          description: >
            When true and this is the tenant's default provider, every alert
            ingested via webhook is analyzed automatically. Off by default:
            analysis then only happens when an analyst clicks "Analyze with AI".
        createdBy: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SaveLLMProviderRequest:
      type: object
      properties:
        name: { type: string }
        kind: { type: string }
        baseUrl: { type: string, nullable: true }
        model: { type: string }
        apiKey:
          type: string
          description: Write-only. Required on create; empty on update means "keep the existing key".
        autoAnalyzeAllAlerts:
          type: boolean
          description: Fully replaces the stored value on every save, unlike apiKey -- there's no "omit to keep existing" case.
      required: [name, model]
...
    RetentionConfig:
      type: object
      description: >
        GET never returns null -- configured=false means these are the
        synthesized 18-month defaults, not a value ever actually saved.
      properties:
        tenantId: { type: string, format: uuid }
        alertRetentionMonths: { type: integer, minimum: 0 }
        incidentRetentionMonths: { type: integer, minimum: 0 }
        configured: { type: boolean }
        updatedAt: { type: string, format: date-time, nullable: true }
...